+++ NEWS.md
@@ -8,6 +8,8 @@
 - Added support for VHDL-2019 sequential block statements.
 - Implemented the VHDL-2019 directory I/O functions in `std.env`.
 - Added VHDL-2019 assert API (with @Blebowski).
+- Implemented `'image`, `'value` and `to_string` for composite types in
+  VHDL-2019.
 
 ## Version 1.10.1 - 2023-07-28
 - Fixed incorrect sensitivity list generation with concurrent statements